Initial check in
Breaking Changes
- The cloning version of the constructor and AddRange() method have been
removed from the following collections.
- DateTimeInstanceCollection
- DayInstanceCollection
- ExtendedBindingList`1 and all types derived from it (various iCalendar
object collections)
- HolidayCollection
- PeriodCollection
- VTimeZoneCollection

It was confusing and inconsistent to have one version that cloned the
range of elements and another that did not. The single AddRange() method
adds the instances from an enumerable list. If you want to add clones,
you must pass an enumerable list of cloned instances to the method now
or use the new CloneRange method on the target collection. The
vCardBrowser and CalendarBrowser demos have been updated to show
examples of its use.

- The HolidayCollection.HolidayList method has been changed to
HolidaysBetween and the return type changed to IEnumerable<DateTime>.
This allows you to create a more efficient collection from the result
set such as a HashSet<DateTime> or to sort the results in a different
way without incurring any overhead from the method.
- HolidayCollection.AddStandardHolidays has been changed to return a
reference to the collection rather than void.
- HolidayManager.Defaults has been change to be gettable and settable
using an IEnumerable<Holiday> value.
- The HolidayManager.GetHolidays() and HolidayManager.SetHolidays()
methods have been removed. Use the new Holidays property instead.
- The RecurOptsDataSource methods are now properties and the underlying
type returned by them was changed from ArrayList to List<ListItem>.
- Made the following changes to the RecurrencePattern web control to
allow for better customization:
- Simplified the control HTML and styling.
- Removed the RenderDefaultStyleSheet property and added a set of
properties to allow setting the CSS class name(s) for the affected
elements: CssButton, CssErrorMessage, CssInput, CssPanel, CssPanelBody,
CssPanelHeading, CssPanelTitle, CssSelect.  This allows the
specification of one or more CSS class names to apply and allows for
using the styles from frameworks such as Bootstrap. The PDIWebDemoCS and
PDIWebDemoVB demo applications contain examples of their use with
Bootstrap.

Other Changes in This Release
- Cleaned up the code and rebuilt the projects for use with .NET 4.0 and
later.
- Updated all collection constructors and AddRange() methods to take
IEnumerable<T> as the parameter type for the items to add. This makes
them more generic and gets rid of the extra methods that used to take an
array or type-specific collection as the parameter.
- All methods in the EncodingUtils class are now extension methods.
- Fixed the RecurrencePattern web control so that it does not try to
parse inner text as child controls.
-  Added a ShowWeekStartDay property to both the Windows Forms and web
control versions of the RecurrencePattern control to allow hiding the
Week Start Day combo box. The layout was changed to place the Week Start
Day and Can Occur on Holiday options on the right to allow for hiding
them both without having to adjust the control layout.
- Converted the project from shareware to open source and released it on
GitHub.